// calendar/src/ui/theme/theme.enum.ts
export enum ThemeColor {
  Primary = "primary",
  Secondary = "secondary",
  Success = "success",
  Danger = "danger",
  Neutral = "neutral",
}

export enum ThemeSize {
  Sm = "sm",
  Md = "md",
  Lg = "lg",
}

export enum ThemeRadius {
  Sm = "sm",
  Base = "base",
  Lg = "lg",
  Xl = "xl",
  Full = "full",
}

export enum ThemeState {
  Default = "default",
  Disabled = "disabled",
  Error = "error",
  Focus = "focus",
  Transition = "transition",
}